% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{addPackage}
\alias{addPackage}
\title{Add package to a CRAN-like repo}
\usage{
addPackage(repopath, bundle, binary = TRUE, oldeRversion = -1L,
  overwrite = FALSE, testbuild = TRUE, multiarch = TRUE)
}
\arguments{
\item{repopath}{path to the repo}

\item{bundle}{package bundle (tar.gz)}

\item{binary}{logical, whether to put the binary source in the repo}

\item{overwrite}{logical, whether to overwrite if the package bundle already exists}

\item{testbuild}{logical, whether to test the package can be built}

\item{multiarch}{logical, whether to test 32bit and 64bit}
}
\value{
Nothing, just prints some messages.
}
\description{
Add a single package to an existing CRAN-like repo.
}
\seealso{
\code{\link{writePACKAGESandArchive}} to scan a repo and update the repo.
}
