% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{archivePackages}
\alias{archivePackages}
\title{Archive packages}
\usage{
archivePackages(repopath, pkgs, type = "source")
}
\arguments{
\item{repopath}{path to the repo}

\item{pkgs}{the packages to be archived}

\item{type}{type of the packages to archive, one of \code{"source"} or \code{"binary"}}
}
\value{
An invisible logical, indicating whether archiving has occured.
}
\description{
Archive older versions of packages found in a repo.
}
\note{
This function is rather for internal use and is not exported.
}
