% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{charseq}
\alias{charseq}
\title{Generates 001, 002, ...}
\usage{
charseq(I, prefix = NULL, suffix = NULL, ndigits = floor(log10(I)) +
  1)
}
\arguments{
\item{I}{integer}

\item{prefix}{character, to prepend}

\item{suffix}{character, to append}

\item{ndigits}{integer, the number of digits; default to no more digits than necessary}
}
\value{
A character vector, see examples. An attribute \code{"string"} is attached, it is used in \code{tikz2png}.
}
\description{
Generates consecutive integers with a given number of digits, with an optional prefix and suffix.
}
\examples{
charseq(11)
charseq(11, ndigits=3)
charseq(11, prefix="graph", suffix=".png")
}
