% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{coltypes}
\alias{coltypes}
\title{Column types}
\usage{
coltypes(dat)
}
\arguments{
\item{dat}{dataframe}
}
\value{
A dataframe with three columns: \code{column}, the column names; \code{class}, the classes of the columns; \code{numerizable} (logical), whether the columns is numerizable.
}
\description{
Returns the types of the columns of a dataframe, and indicates which columns can be faithfully converted
to numeric mode.
}
\note{
This functions uses \code{\link{isNumerizable.df}}.
}
\examples{
dat <- data.frame(x=1:4,
                  y=c("a","b","c","d"),
                  z=c("1", "1.4", "-3", "   0"))
coltypes(dat)

}
