% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactPredictionInterval.R
\name{CoverageEPI}
\alias{CoverageEPI}
\title{Coverage probability of Dunnett's exact prediction intervals}
\usage{
CoverageEPI(nsims, type = "two-sided", conf, n, k)
}
\arguments{
\item{nsims}{positive integer, the number of simulations}

\item{type}{character, type of the interval}

\item{conf}{vector of numbers between 0 and 1, the confidence levels}

\item{n}{vector of integers greater than 1, the sample sizes}

\item{k}{vector of integers greater than 2, the numbers of future
observations}
}
\value{
A datatable.
}
\description{
Estimates the coverage probability of Dunnett's
exact prediction intervals with the help of simulations
}
\examples{
CoverageEPI(10000, conf=c(0.8, 0.9), n=c(5,10), k=c(2,3))
}
