% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{createFolder}
\alias{createFolder}
\title{Create a new folder}
\usage{
createFolder(folder, overwrite = FALSE)
}
\arguments{
\item{folder}{path to the folder to create (absolute or relative)}

\item{overwrite}{logical, whether to overwrite if the folder already exists}
}
\value{
The path of the created folder, if it is created.
}
\description{
Create a new folder.
}
