% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{datify}
\alias{datify}
\title{Format dates for LaTeX}
\usage{
datify()
}
\description{
Format dates to be used in LaTeX with \code{\\usepackage[ddmmyyyy]{datetime}}.
}
\details{
This function is useful to be used in the \code{labels} argument of the \code{ggplot2} function \code{\link[ggplot2]{scale_x_date}} (see the example).
}
\note{
See \url{http://stackoverflow.com/questions/20326946/how-to-put-ggplot2-ticks-labels-between-dollars} for possible additional choices of the format.
}
\examples{
datify()(Sys.Date() + 0:1)
\dontrun{
# a base plot example
plotCode <- function(){
  par(mar=c(7,4,4,2)+0.1)
  plot(1:8, pch=19, xaxt="n", yaxt="n", xlab=NA, ylab="$y$")
  axis(2, at=1:8, labels=dollarify()(1:8))
  axis(1, labels=FALSE)
  labels <- datify()(Sys.Date()-1:8)
  text(1:8, par("usr")[3] - 0.5, srt=45, adj=1,
       labels=labels, xpd=TRUE)
  mtext(1, text ="\\\\textsf{date}", line = 5)
}
# create TikZ plot
plot2tikz(plotCode, compile=FALSE, outdir=tempdir(),  overwrite = TRUE,
          packages=c("\\\\usepackage[ddmmyyyy]{datetime}\\n"),
          documentDeclaration ="\\\\documentclass[12pt]{standalone}\\n",
          width=7, height=5)}
\dontrun{
# a ggplot2 example
library(ggplot2)
DF <- data.frame(date=Sys.Date()+1:20, y=rpois(20, lambda=30))
gg <- ggplot(DF, aes(x=date, y=y)) +
  geom_point(size=4, color="yellow", shape=21, fill="green") +
  scale_y_continuous(labels=dollarify()) +
  scale_x_date(date_breaks="2 days", labels=datify()) +
  xlab("\\\\large\\\\textsf{date}") +
  ylab("\\\\Large $y$") +
  theme(axis.text.x = element_text(face="bold", angle=45, vjust=0.5, size=12, lineheight=0.9),
        axis.text.y = element_text(face="bold", size=16, lineheight=0.9, hjust=1),
        axis.ticks = element_line(size=0.2),
        axis.title.x = element_text(size=20, vjust=0),
        axis.title.y = element_text(face="bold", size=16, vjust=1),
        axis.ticks.length = unit(0.3, "lines"),
        axis.text = element_text(margin=unit(0.5, "lines"))
  )
# create TikZ plot
plot2tikz(function() print(gg), compile=FALSE, outdir=tempdir(), overwrite = TRUE,
          packages=c("\\\\usepackage[ddmmyyyy]{datetime}\\n"),
          documentDeclaration ="\\\\documentclass[12pt]{standalone}\\n",
          width=7, height=5)}
}
