% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{dollarify}
\alias{dollarify}
\title{Enclose between dollars}
\usage{
dollarify()
}
\value{
A function surrounding its input between \code{$}'s.
}
\description{
Returns a function surrounding its input between \code{$}'s.
}
\details{
This function is useful to be used in the \code{labels} argument of the \code{ggplot2} function \code{scale_[x/y]_continuous} (see the example of \code{\link{datify}}).
}
\note{
This function is useful for using \code{\link{plot2tikz}}.
}
\examples{
curve(x^2, from=-1, to=1, asp=1, axes=FALSE, xlab="$x$", ylab=NA)
xlabs_at <- seq(-1, 1, by=0.5)
axis(1, at=xlabs_at, labels=dollarify()(xlabs_at))
ylabs_at <- seq(0, 1, by=0.2)
axis(2, at=ylabs_at, labels=dollarify()(ylabs_at))
text(0.65, 0.8, "$f(x)=x\\\\\\\\^2$") # note the four backslashes!
}
