% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dunnett.R
\name{DunnettDistribution}
\alias{DunnettDistribution}
\alias{qDunnett}
\alias{pDunnett}
\alias{rDunnett}
\title{Cumulative distribution function and quantile function of the
Dunnett distribution}
\usage{
qDunnett(p, df, k, rho, tailed = c("one", "two"),
  tol = .Machine$double.eps, maxiter = 5000)

pDunnett(q, df, k, rho, tailed = c("one", "two"))

rDunnett(n, df, k, rho, tailed = c("one", "two"))
}
\arguments{
\item{p}{number between 0 and 1, the probability to achieve}

\item{df}{postive number, the degrees of freedom}

\item{k}{integer greater than or equal to 2, the dimension}

\item{rho}{number between 0 and 1, correlation}

\item{tailed}{\code{"one"} or \code{"two"}, for the one-tailed Dunnett
dsitribution and the two-tailed Dunnett distribution, respectively}

\item{tol, maxiter}{parameters passed to \code{\link[stats]{uniroot}}}

\item{q}{numeric, the quantile}

\item{n}{number of simulations}
}
\value{
\code{pDunnett} returns the value of the cdf, \code{qDunnett}
returns the value of the inverse cdf, \code{rDunnett} samples the Dunnett
distribution.
}
\description{
Cumulative distribution function and quantile function of
the Dunnett distribution.
}
\details{
See the vignette.
}
\examples{
qDunnett(0.95, df=3, k=2, rho=0.5)
pDunnett(3.87, df=3, k=2, rho=0.5)
}
