% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{formatClatex}
\alias{formatClatex}
\title{Formatting a number for LaTeX}
\usage{
formatClatex(x, ...)
}
\arguments{
\item{x}{number}

\item{...}{arguments passed to \code{\link[base]{formatC}}}
}
\value{
A character string ready to be surrounded between dollars for LaTeX.
}
\description{
Formatting a number for LaTeX, handling the scientific notation
}
\examples{
formatClatex(1.234567, digits=4)
formatClatex(1.234567e-12, digits=4)
}
