% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl.R
\name{getExcelSheet}
\alias{getExcelSheet}
\title{Get an Excel sheet}
\usage{
getExcelSheet(xlsxFile, sheet = 1L)
}
\arguments{
\item{xlsxFile}{path to a xlsx file}

\item{sheet}{sheet index or name}
}
\value{
A keyed datatable.
}
\description{
Get the contents of an Excel sheet cell by cell.
}
\note{
This function is intended to be used for \code{\link{sheetToJSON}} and \code{\link{treadXLSX}}.
}
\examples{
xlsxFile <- system.file("extdata", "xlsx", "xlsx_small.xlsx", package="SLutils")
sheet <- getExcelSheet(xlsxFile, sheet="Sheet2")
library(data.table)
key(sheet)
sheet[.(2,4), .(data_type, character, comment)]
}
