% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{guessDateFormat}
\alias{guessDateFormat}
\title{Guess date format}
\usage{
guessDateFormat(x, returnDates = FALSE, tzone = "", messages = FALSE)
}
\arguments{
\item{x}{a vector (not a factor)}

\item{returnDates}{logical, whether to return the dates}

\item{tzone}{time zone}
}
\description{
Guess the date format.
}
\author{
Cole Beck.
}
