% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable.R
\name{hotColumnsFromTypes}
\alias{hotColumnsFromTypes}
\title{Create handsontable columns setting from given types}
\usage{
hotColumnsFromTypes(dftypes)
}
\arguments{
\item{dftypes}{dataframe as returned by \code{\link{XLSXtypes}}}
}
