% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{hreadXLSX}
\alias{hreadXLSX}
\title{Read a XLSX file}
\usage{
hreadXLSX(file, sheet = "Sheet1", what = "data", header = TRUE,
  firstRow = NULL, lastRow = NULL, output = c("flatlist", "list",
  "dataframe", "JSON"), stringsAsFactors = TRUE)
}
\arguments{
\item{file}{name of the output xlsx file}

\item{sheet}{sheet name}

\item{what}{what to read; must be one among \code{"data"}, \code{"comments"},
\code{"types"}, \code{"formats"}, or comma-separated combination of these choices (e.g. "data,comments")}

\item{header}{whether to consider the first row as column headers}

\item{firstRow}{integer or \code{NULL}, the starting row (guessed if \code{NULL})}

\item{lastRow}{integer or \code{NULL}, the ending row (guessed if \code{NULL})}

\item{output}{the desired output; see \strong{Value}}
}
\value{
The possible output is:
\itemize{
     \item If only one "what" is requested:
     a "flat list" if \code{output="flatlist"},
a list if \code{output="list"}, a dataframe if \code{output="dataframe"},
a JSON string if \code{output="JSON"}.
     \item Otherwise, a list of the previous kind of outputs.
}
}
\description{
Read a XLSX file
}
\examples{
xlsx <- system.file("extdata/xlsx/xlsx_comments.xlsx", package = "SLutils")
hreadXLSX(xlsx, sheet="Sheet1")
}
