% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install.package.vanilla}
\alias{install.package.vanilla}
\title{Install a package in "vanilla" mode}
\usage{
install.package.vanilla(pkgSource, lib)
}
\arguments{
\item{pkgSource}{the path (absolute or relative) to the tar.gz bundle}

\item{lib}{character, the library directory where to install the package; if missing, defaults to the first element of \code{.libPaths()}}
}
\value{
The output of the system command.
}
\description{
Install a package with option \code{--vanilla}.
}
