% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install.packages.vanilla2}
\alias{install.packages.vanilla2}
\title{Install some packages in "vanilla" mode}
\usage{
install.packages.vanilla2(pkgs, ...)
}
\arguments{
\item{pkgs}{vector of packages to be installed}

\item{...}{passed to \code{\link[utils]{install.packages}}}
}
\description{
Install some packages with option \code{--vanilla}.
}
\examples{
\dontrun{install.packages.vanilla2("abind"))}
}
