% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{isNumerizable.df}
\alias{isNumerizable.df}
\title{Test whether columns of a dataframe are numerizable}
\usage{
isNumerizable.df(dat)
}
\arguments{
\item{dat}{dataframe}
}
\value{
A named logical vector, indicating for each column whether this column is numerizable.
}
\description{
Tests which columns of a dataframe can be faithfully converted to a numeric vector.
}
\note{
This function is based on \code{\link{isNumerizable}}.
}
\examples{
dat <- data.frame(x=1:4,
                  y=c("a","b","c","d"),
                  z=c("1", "1.4", "-3", "   0"))
isNumerizable.df(dat)

}
\seealso{
\code{\link{coltypes}}, which returns the types of the columns of a dataframe, and indicates which ones are numerizable.
}
