% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{isNumerizable}
\alias{isNumerizable}
\title{Test whether a vector contains numbers only}
\usage{
isNumerizable(string)
}
\arguments{
\item{string}{a vector or a factor}
}
\value{
A logical value. When the result is \code{FALSE}, an attribute \code{"unnumerizable"} is set to this value, it indicates which elements are not numerizable.
}
\description{
Test whether a vector contains numbers only and then can be faithfully converted to a numeric vector.
}
\note{
When the result is \code{TRUE}, then \code{string} can be faithfully converted to a numeric vector with \code{\link{Numerize}}.
\strong{Warning:} If \code{string} is a factor and the result is \code{TRUE}, then \code{as.numeric(string)} does not convert to the expected numeric vector (use \code{Numerize(string)} or \code{as.numeric(as.character(string))}).
}
\examples{
isNumerizable(c(1, 2, 3))
isNumerizable(c("a","b","c"))
isNumerizable(c("1", "2", "3"))
isNumerizable(c("1", "b", "c"))
isNumerizable(c("1.5", "  3.2 "))
isNumerizable(factor(c("3.5", "4")))

}
\seealso{
\code{\link{Numerize}}, which is based on \code{isNumerizable}.
}
