% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{makeArchiveRDS}
\alias{makeArchiveRDS}
\title{Make the archive.rds file of a repo}
\usage{
makeArchiveRDS(repopath, type = "source")
}
\arguments{
\item{repopath}{path to the repo}

\item{type}{type of the packages to archive, one of \code{"source"} or \code{"binary"}}
}
\value{
Nothing.
}
\description{
Make the archive.rds file of a CRAN-like repo.
}
\note{
This function is rather for internal use and is not exported.
}
