% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{makeRepo}
\alias{makeRepo}
\title{Create a CRAN-like repo}
\usage{
makeRepo(repopath, name, overwrite = FALSE)
}
\arguments{
\item{repopath}{path to the repo (absolute or relative)}

\item{name}{name of the repo}

\item{overwrite}{logical, whether to overwrite if the repo already exists}
}
\value{
The full path of the repo.
}
\description{
Create a CRAN-like repo
}
