% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{blockmatrices}
\alias{blockmatrices}
\alias{blockdiag}
\alias{blockdiag_list}
\title{Create block diagonal matrix}
\usage{
blockdiag(M1, M2)

blockdiag_list(Ms)
}
\arguments{
\item{M1}{top-left block}

\item{M2}{bottom-right block}

\item{Ms}{list of matrices}
}
\value{
A matrix.
}
\description{
Build a block diagonal matrix
}
\examples{
M1 <- diag(2); M2 <- matrix(2, nrow=3, ncol=3)
blockdiag(M1,M2)
blockdiag_list(list(M1,M2,M1))
}
\seealso{
The \code{\link[Matrix]{bdiag}} function in the \code{Matrix} package.
}
