% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{ndecimals}
\alias{ndecimals}
\title{Number of decimals}
\usage{
ndecimals(x)
}
\arguments{
\item{x}{number}
}
\value{
The number of decimals of \code{x}.
}
\description{
Number of decimals of a number.
}
\examples{
ndecimals(13.451)
ndecimals(2)
ndecimals(-43.421)
ndecimals(pi)
ndecimals(1.23e-20)
}
\seealso{
\code{\link{nfloordigits}}
}
