% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{nfloordigits}
\alias{nfloordigits}
\title{Number of digits of the integer part}
\usage{
nfloordigits(x)
}
\arguments{
\item{x}{number}
}
\value{
The number of digits of the integer part of \code{x}.
}
\description{
Number of digits of the integer part of a number.
}
\examples{
nfloordigits(13.451)
nfloordigits(2)
nfloordigits(-43.421)
nfloordigits(23.15e6)
}
\seealso{
\code{\link{ndecimals}}
}
