% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert numbers to words}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
The vector of words.
}
\description{
Convert numbers to their corresponding English words.
}
\examples{
numbers2words(c(2, 17))
}
\author{
\href{http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html}{John Fox}
and \href{https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r}{Tony Hirst}
}
