% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{Numerize.df}
\alias{Numerize.df}
\title{Numerize a dataframe}
\usage{
Numerize.df(dat, factors = TRUE, colnames = names(dat))
}
\arguments{
\item{dat}{a dataframe whose is each column class is character or factor}

\item{factors}{logical indicating to include factor columns too}

\item{colnames}{the names of the columns to (attempt to) numerize}
}
\value{
A dataframe, the same as \code{dat} after transforming its columns to numeric mode when possible. When some columns are not numerizable, their names are given in the attribute \code{"unnumerizable"}.
}
\description{
Given a dataframe of character columns, returns the same dataframe with numeric columns when possible.
}
\note{
This function is based on \code{\link{isNumerizable}}.
}
\examples{
dat <- data.frame(x=1:4,
                  y=c("a","b","c","d"),
                  z=c("1", "1.4", "-3", "   0"))
( numdat <- Numerize.df(dat) )
attr(numdat, "unnumerizable")

}
\seealso{
\code{\link{Numerize}} to numerize a vector.
}
