% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{Numerize}
\alias{Numerize}
\title{Numerize a character/factor vector}
\usage{
Numerize(x)
}
\arguments{
\item{x}{character or factor vector to be numerized}
}
\value{
A numeric vector if possible, otherwise an error.
}
\description{
Given a character or factor vector, returns the same in numeric mode if possible.
}
\note{
This function is based on \code{\link{isNumerizable}}.
}
\examples{
Numerize(c("1", "1.4", "-3", "   0"))
\dontrun{
Numerize(c("a","b","c","d"))}

}
\seealso{
\code{\link{Numerize.df}} to numerize a dataframe, and the function of the same kind \code{\link[utils]{type.convert}}.
}
