% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{numextract}
\alias{numextract}
\title{Numbers extraction - first number only}
\usage{
numextract(string, allow.negative = TRUE)
}
\arguments{
\item{string}{a character vector}

\item{allow.negative}{logical, whether to allow negative numbers, or to ignore the minus sign}
}
\value{
A character vector with the same length as \code{string}. A \code{NA} is returned when there is no number to extract from the character string.
}
\description{
Extract the first number contained in a character string.
}
\note{
If the character string contains several numbers, this function extracts the first one only.
}
\examples{
numextract("30.5ml")
numextract(">2g")
numextract("between 5 and 6")
numextract(c("30.5ml", "37ml"))
numextract("-3C")
numextract("-3C", allow.negative=FALSE)
numextract(c("1", "b"))
numextract(c("1", NA))
numextract(c("1", "NA"))

}
\seealso{
\code{\link{numsextract}} to extract all numbers in a character string, \code{\link{isNumerizable}}, which is based on \code{numextract}, and the similar function \code{readr::parse_number}.
}
