% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numextract.R
\name{numsextract}
\alias{numsextract}
\title{Numbers extraction}
\usage{
numsextract(string, allow.negative = TRUE)
}
\arguments{
\item{string}{a character string}

\item{allow.negative}{logical, whether to allow negative numbers, or to ignore the minus sign}
}
\value{
A character vector, the numbers contained in \code{string}.
}
\description{
Extract all numbers contained in a character string.
}
\examples{
numsextract("between 2ml and 3.5ml")
numsextract("between -4.1 and -2")
numsextract("between -4.1 and -2", allow.negative=FALSE)

}
\seealso{
\code{\link{numextract}} to extract the first number only.
}
