% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{plot2tikz}
\alias{plot2tikz}
\title{Convert a plot to a TikZ figure}
\usage{
plot2tikz(code, filename = "Rplot", outdir = getwd(),
  overwrite = FALSE, format = "pdf", lua = FALSE, packages = NULL,
  addDefaultTikZoptions = TRUE, compile = TRUE, clean = FALSE, ...)
}
\arguments{
\item{code}{a function without argument generating the plot}

\item{filename}{the name of the tex file, without extension}

\item{outdir}{output directory}

\item{overwrite}{logical, whether to overwrite if the file exists}

\item{format}{the output format, one of \code{"pdf"}, \code{"ps"}, or \code{"eps"}}

\item{lua}{logical, wheter to use \code{lualatex}}

\item{packages}{\code{NULL} to use default packages, otherwise see examples}

\item{addDefaultTikZoptions}{logical; for \code{format="pdf"} only and if \code{packages} is not \code{NULL}, then this adds the default packages given in \code{getOption("tikzLatexPackages")}}

\item{compile}{logical, whether to compile the tex file; automatically \code{TRUE} if \code{format="eps"}}

\item{clean}{logical, whether to remove the auxiliary files}

\item{...}{arguments passed to \code{\link[tikzDevice]{tikz}}, such as \code{documentDeclaration}, \code{width} and \code{height}.}
}
\value{
No value, only prints some messages.
}
\description{
Convert a R plot to a TikZ figure
}
\examples{
\dontrun{
# first write a function creating the plot
plotCode <- function(){
  curve(x^2, from=-1, to=1, asp=1, axes=FALSE, xlab="$x$", ylab=NA)
  xlabs_at <- seq(-1, 1, by=0.5)
  axis(1, at=xlabs_at, labels=dollarify()(xlabs_at))
  ylabs_at <- seq(0, 1, by=0.2)
  axis(2, at=ylabs_at, labels=dollarify()(ylabs_at))
  text(0.65, 0.8, "$f(x)=x\\\\\\\\^2$") # note the four backslashes!
  return(invisible())
}
# try the plot:
plotCode()
# create the LaTeX file with default preamble for pdflatex
plot2tikz(plotCode, outdir=tempdir(), compile=FALSE,
  documentDeclaration ="\\\\documentclass[12pt]{standalone}\\n",
  width=7, height=5)
# add some packages, or other commands in the preamble
plot2tikz(plotCode, compile=FALSE, outdir=tempdir(), overwrite = TRUE,
          packages=c("\\\\usepackage[active,tightpage,psfixbb]{preview}\\n",
                     "\\\\PreviewEnvironment{pgfpicture}\\n",
                     "\\\\setlength\\\\PreviewBorder{10pt}\\n",
                     "\\\\usepackage{amssymb}\\n"),
          documentDeclaration ="\\\\documentclass[12pt]{standalone}\\n",
          width=7, height=5)}
}
\seealso{
Examples in \code{\link{datify}}.
}
