% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactPredictionInterval.R
\name{predictionInterval}
\alias{predictionInterval}
\title{Exact prediction interval for a Gaussian sample}
\usage{
predictionInterval(x, n.mean = 1, k = 1, method = "Bonferroni",
  type = "two-sided", conf.level = 0.95)
}
\arguments{
\item{x}{sample}

\item{n.mean}{positive integer, the sample size associated with the \code{k}
future averages; the default value \code{n.mean=1} means individual observations;
note that all future averages must be based on the same sample size}

\item{k}{positive integer, the number of future observations or averages
the prediction interval should contain with the given confidence level}

\item{method}{character string, the method to use if the number \code{k} of
future observations is greater than 1 (other this argument is ignored);
possible values are \code{"Bonferroni"} (approximate method based on the
Bonferonni inequality) and  \code{"exact"} (exact method due to Dunnett, 1955;
see the \strong{Details} section of \code{\link[EnvStats]{predIntNormK}} for
more information.}

\item{type}{interval type, \code{"two-sided"}, \code{"lower"} or \code{"upper"}}

\item{conf.level}{confidence level}
}
\value{
A numeric vector of length two if \code{"two-sided"}, otherwise the bound
in the one-sided case.
}
\description{
Exact prediction interval for the next \code{k} observations
or next set of means. This is a port of the function \code{\link[EnvStats]{predIntNormK}}
od the \code{EnvStats} package.
}
\details{
See the vignette.
}
\examples{
predictionInterval(rnorm(10), k=3, method="exact")
}
