% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{prettyNumber}
\alias{prettyNumber}
\title{Format a number with a given number of characters}
\usage{
prettyNumber(x, digits = 9L)
}
\arguments{
\item{x}{number}

\item{digits}{maximal number of characters}
}
\description{
Format (and round) a number with a given maximal number of characters.
}
\examples{
prettyNumber(pi, 4)
prettyNumber(123456789, 7)
prettyNumber(12.5e9, 7)
prettyNumber(0.00000012, 7)
}
