% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pspdf2img}
\alias{pspdf2img}
\title{Convert (e)ps to png/gif}
\usage{
pspdf2img(infile, format = "png", moreopts = "", ps2eps = FALSE)
}
\arguments{
\item{infile}{input file}

\item{format}{output format}

\item{moreopts}{more options to \code{convert} of ImageMagick}

\item{ps2eps}{apply \code{ps2eps} before converting, for a ps input file}
}
\value{
Nothing, this only creates the file.
}
\description{
Convert a (e)ps file to png or gif
}
\examples{
\dontrun{
pdf("Rplot.pdf")
curve(x^2, from=-1, to=1, axes=FALSE, ylab=NA, col="red")
xlabs_at <- seq(-1, 1, by=0.5)
axis(1, at=xlabs_at)
ylabs_at <- seq(0, 1, by=0.2)
axis(2, at=ylabs_at)
text(0.65, 0.8, expression(f(x)==x^2))
dev.off()
pspdf2img("Rplot.pdf", format="png", moreopts="-trim -resize 50\%")}
}
