% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{repoStatus}
\alias{repoStatus}
\title{Get packages versions in a repo}
\usage{
repoStatus(repopath, pkgs, type = "source")
}
\arguments{
\item{repopath}{path to the repo}

\item{pkgs}{the packages to look for}

\item{type}{type of the packages to archive, one of \code{"source"} or \code{"binary"}}
}
\value{
The packages and their version in a dataframe.
}
\description{
Return the packages in a repo and their version.
}
\note{
This function is rather for internal use and is not exported.
}
