% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javascript.R
\name{scianim}
\alias{scianim}
\title{Create html and js files for Scianimator}
\usage{
scianim(name, imgs, theme = "", outdir = getwd())
}
\arguments{
\item{name}{name of the html file (without extension)}

\item{imgs}{a list as the output of \code{\link{tikz2png}}: a field \code{files} containg the paths of the files, and a field \code{json} containing a JSON string}

\item{theme}{the theme, one of \code{""}, \code{"light"}, \code{"blue"}, \code{"dark"}}

\item{outdir}{output directory}
}
\value{
No value. This creates the html file and its dependencies.
}
\description{
Create html and js files for Scianimator
}
