% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl.R
\name{sheetToJSON}
\alias{sheetToJSON}
\title{Excel sheet to JSON strings}
\usage{
sheetToJSON(sheet, header = TRUE, firstRow = NULL)
}
\arguments{
\item{sheet}{an Excel sheet, like the output of \code{\link{getExcelSheet}}}

\item{header}{logical, whether considering the first row as column headers}

\item{firstRow}{the first row to be read, an integer or \code{NULL} to guess it}
}
\value{
A list of three JSON strings.
}
\description{
Convert an Excel sheet (output of \code{\link{getExcelSheet}}) to
a list of three JSON strings: values, comments, and types.
}
\note{
This function is a shortcut to \code{\link{treadXLSX}} with option
\code{output="JSON"}.
}
\examples{
xlsxFile <- system.file("extdata", "xlsx", "xlsx_mixed.xlsx", package="SLutils")
sheet <- getExcelSheet(xlsxFile)
sheetToJSON(sheet)
sheetToJSON(sheet, header=FALSE)$data
sheetToJSON(sheet, header=FALSE, firstRow=3)$data
}
\seealso{
\code{\link{getExcelSheet}}, \code{\link{treadXLSX}}.
}
