% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string2letters}
\alias{string2letters}
\title{Separate characters of a string}
\usage{
string2letters(string)
}
\arguments{
\item{string}{a character string}
}
\value{
A character vector.
}
\description{
Separate the characters of a character string.
}
\examples{
# this returns c("h", "e", "l", "l", "o"):
string2letters("hello")
}
