% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Student.R
\name{StudentDistribution}
\alias{StudentDistribution}
\alias{pT}
\alias{qT}
\title{Cumulative distribution function and quantile function of the Student t distribution}
\usage{
pT(q, df, ncp = 0, lower.tail = TRUE, log.p = FALSE, max.terms = 1e+05,
  num.terms.inc = 20, rel.tol = .Machine$double.eps)

qT(p, df, ncp = 0, lower.tail = TRUE, log.p = FALSE, max.terms = 1e+05,
  num.terms.inc = 20, rel.tol = .Machine$double.eps)
}
\arguments{
\item{q}{vector of quantiles}

\item{df}{degrees of freedom}

\item{ncp}{non-centrality parameter}

\item{p}{vector of probabilities}
}
\value{
\code{pT} returns the value(s) of the cdf, \code{qT} returns the value(s) of the quantile function.
}
\description{
Cumulative distribution function and quantile function of the central or noncentral Student t distribution.
}
\note{
These functions are copied from the \code{EnvStats} package.
They give a correct result when the non-centrality parameter is large while \code{\link[stats]{pt}} or \code{\link[stats]{qt}} fails.
}
