% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{subfoldersCRANlike}
\alias{subfoldersCRANlike}
\title{Paths of the CRAN-like subfolders
Return the structure of a CRAN-like repo.}
\usage{
subfoldersCRANlike(repopath)
}
\arguments{
\item{repopath}{path to the repo (not necessarily an existing one)}
}
\value{
The structure of the repo.
}
\description{
Paths of the CRAN-like subfolders
Return the structure of a CRAN-like repo.
}
\note{
This function is rather for internal use and is not exported.
}
