% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Sys.which2}
\alias{Sys.which2}
\title{A modification of \code{Sys.which}}
\usage{
Sys.which2(cmd)
}
\arguments{
\item{cmd}{name of possible executable to search}
}
\value{
The path of the executable, if found, otherwise the empty string.
}
\description{
Find an executable without including the Windows system directories before \code{PATH}.
}
\note{
Credit to \href{http://stackoverflow.com/q/34030087/1100107}{Henrik Bengtsson}.
}
\examples{
Sys.which("convert")
Sys.which2("convert")
}
