% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tikz2png}
\alias{tikz2png}
\title{Generate LaTeX figures in png or gif format}
\usage{
tikz2png(code, nplots, prefix = "Rplot", format = "png",
  outdir = getwd(), folder = "assets/images/", ...)
}
\arguments{
\item{code}{a function with an integer argument, generating the plot}

\item{nplots}{an integer vector, each value is applied to the \code{code} function}

\item{prefix}{the prefix for the generated files}

\item{format}{the desired output format, \code{"png"} or \code{"gif"}}

\item{outdir}{output directory}

\item{folder}{this is intended to be used with \code{\link{scianim}}}

\item{...}{arguments passed to \code{\link[tikzDevice]{tikz}}}
}
\value{
A list with two components, the paths of the generated files and a JSON string intended to be used with \code{\link{scianim}}.
}
\description{
Generate a series of TikZ figures in png or gif format
}
\examples{
\dontrun{
plotCode <- function(n){
  curve(x^n, from=-1, to=1, axes=FALSE,
        ylab=NA, xlab="$x$",
        main=sprintf("$f(x)=x\\\\\\\\^\%s$", n))
  xlabs_at <- seq(-1, 1, by=0.5)
  axis(1, at=xlabs_at, labels=dollarify()(xlabs_at))
  ylabs_at <- seq(ifelse(n\%\%2==0, 0, -1), 1, by=0.2)
  axis(2, at=ylabs_at, labels=dollarify()(ylabs_at))
  return(invisible())
}
tikz2png(plotCode, 1:3, outdir = tempdir())}
}
