% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl.R
\name{treadXLSX}
\alias{treadXLSX}
\title{Read an Excel sheet}
\usage{
treadXLSX(sheet, header = TRUE, firstRow = NULL, output = c("list",
  "JSON", "both"))
}
\arguments{
\item{sheet}{an Excel sheet, like the output of \code{\link{getExcelSheet}}}

\item{header}{logical, whether considering the first row as column headers}

\item{firstRow}{the first row to be read, an integer or \code{NULL} to guess it}

\item{output}{\code{"list"} to get a list of three R objects
(\code{"data"}: the list of columns in the worksheet,
\code{"comments"}: a tibble containing the comments of the cells,
\code{"types"}: a tibble containing the comments of the cells),
\code{"JSON"} to get a list of these three objects encoded to JSON strings,
or \code{"both"} to get both.}
}
\value{
A list of three R objects or three JSON strings, or both.
}
\description{
Convert an Excel sheet (output of \code{\link{getExcelSheet}}) to
a list of three R objects or JSON strings: values, comments, and types.
}
\examples{
xlsxFile <- system.file("extdata", "xlsx", "xlsx_mixed.xlsx", package="SLutils")
sheet <- getExcelSheet(xlsxFile)
treadXLSX(sheet)
identical(treadXLSX(sheet, output="JSON"), sheetToJSON(sheet))
treadXLSX(sheet, header=FALSE)$data
treadXLSX(sheet, header=FALSE, firstRow=3)$data$X2
}
\seealso{
\code{\link{getExcelSheet}}, \code{\link{hreadXLSX}}, a similar function,
and \code{\link{sheetToJSON}}, a shortcut to \code{output="JSON"}.
}
