% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{typeIItable}
\alias{typeIItable}
\title{Type II ANOVA table}
\usage{
typeIItable(fit)
}
\arguments{
\item{fit}{a \code{lm} object, with a formula of the form \code{y~a+b+a:b}}
}
\value{
The table, class \code{anova} and \code{data.frame}.
}
\description{
Type II ANOVA table for a model with two covariates including the interaction
}
\examples{
fit <- lm(yield ~ block + N + block:N, data=npk)
typeIItable(fit)
}
