% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{writePACKAGESandArchive}
\alias{writePACKAGESandArchive}
\title{Write PACKAGES files and archive in a repo.}
\usage{
writePACKAGESandArchive(repopath)
}
\arguments{
\item{repopath}{the path to the repo folder}
}
\value{
Invisibly returns two logical values, indicating whether an archiving occurs for source packages and binary packages.
}
\description{
Scan a CRAN-like repo and update it.
}
\examples{
\dontrun{repopath <- "~/myrepo"
repopath <- makeRepo(repopath)
# put some tarballs in contrib/src ...
writePACKAGESandArchive(repopath)}
}
\seealso{
\code{\link{addPackage}} to add a single package to a repo.
}
