% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx.R
\name{writeXLSX}
\alias{writeXLSX}
\title{Write a XLSX file from a flat list}
\usage{
writeXLSX(flatlist, file, sheetname = "Sheet1", comments = NULL,
  commentsAuthor = NULL, guessDates = TRUE, base64 = FALSE)
}
\arguments{
\item{flatlist}{a list with structure like the output
of \code{\link[jsonlite]{fromJSON}} with option \code{simplifyVector=FALSE}}

\item{file}{name of the output xlsx file}

\item{sheetname}{name of the sheet to be written}

\item{comments}{comments, structure must be similar to \code{flatlist}}

\item{commentsAuthor}{author name of the comments}

\item{guessDates}{whether to guess dates and then write with these cells with date format}

\item{base64}{logical, whether to get the file as a base64 string}
}
\value{
No value if \code{base64=FALSE}, otherwise a base64 string.
}
\description{
Write a XLSX file from a flat list, allowing different types within columns
}
\examples{
\dontrun{
flatlist <- jsonlite::fromJSON("{\\"COL1\\":[1,null,2],\\"COL2\\":[\\"c\\",4],\\"Date\\":[\\"2017-02-12\\"]}", simplifyVector=FALSE)
comments <- jsonlite::fromJSON("{\\"COL1\\":[\\"hello\\",null,null],\\"COL2\\":[null,\\"bye\\"],\\"Date\\":[null]}", simplifyVector=FALSE)
writeXLSX(flatlist, file="xlsx.xlsx", comments=comments)}
}
