% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{XLSXcomments}
\alias{XLSXcomments}
\title{Read the comments in a XLSX file}
\usage{
XLSXcomments(file, sheet = "Sheet1", header = TRUE, firstRow = NULL,
  lastRow = NULL, output = c("dataframe", "JSON"))
}
\arguments{
\item{file}{xlsx file to be read}

\item{sheet}{name of the sheet to be read}

\item{header}{logical, whether the table has column headers}

\item{firstRow}{integer, the starting row (guessed if \code{NULL})}

\item{lastRow}{integer, the ending row (guessed if \code{NULL})}

\item{output}{\code{"dataframe"} or \code{"JSON"}; see \strong{Value}}
}
\value{
A dataframe if \code{output="dataframe"}, a JSON string if \code{output="JSON"}.
}
\description{
Read the comments in the sheet of a XLSX file.
}
\examples{
xlsx <- system.file("extdata/xlsx/xlsx_comments.xlsx", package = "SLutils")
XLSXcomments(xlsx, sheet="Sheet1")
}
