% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{XLSXread}
\alias{XLSXread}
\title{Read a XLSX file}
\usage{
XLSXread(file, sheet = NULL, header = TRUE, comments = FALSE,
  firstRow = NULL, lastRow = NULL, output = c("dataframe", "JSON"))
}
\arguments{
\item{file}{xlsx file to be read}

\item{sheet}{name of the sheet to be read, or \code{NULL} to read all available sheets}

\item{header}{logical, whether the table has column headers}

\item{comments}{logical, whether to read the comments}

\item{firstRow}{integer, the starting row (guessed if \code{NULL}); ignored when \code{sheet=NULL}}

\item{lastRow}{integer, the ending row (guessed if \code{NULL}); ignored when \code{sheet=NULL}}

\item{output}{\code{"dataframe"} or \code{"JSON"}; see \strong{Value}}
}
\value{
If \code{comments=FALSE} and \code{output="dataframe"}: a dataframe if \code{sheet} is not \code{NULL}, a list of dataframes if \code{sheet} is \code{NULL}.
If \code{comments=TRUE} and \code{output="dataframe"}: a list of two dataframes if \code{sheet} is not \code{NULL}, a list of list of two dataframes if \code{sheet} is \code{NULL}.
If \code{output="JSON"}: a JSON string.
}
\description{
Read a sheet or all sheets in a XLSX file.
}
\examples{
\dontrun{
xlsx <- system.file("extdata/datasets.xlsx", package = "readxl")
# try to read an unavailable sheet:
XLSXread(xlsx, sheet="Sheet1")
# read an available sheet:
dat <- XLSXread(xlsx, sheet="iris")
# read all available sheets:
allsheets <- XLSXread(xlsx)
names(allsheets)}
}
