% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{XLSXsheets}
\alias{XLSXsheets}
\title{Sheets in a XLSX file}
\usage{
XLSXsheets(file, output = c("vector", "JSON"))
}
\arguments{
\item{file}{xlsx file to be read}

\item{output}{\code{"vector"} or \code{"JSON"}; see \strong{Value}}
}
\value{
A character vector if \code{output="vector"}, a JSON string if \code{output="JSON"}.
}
\description{
Read the sheet names of a XLSX file.
}
\examples{
xlsx <- system.file("extdata/xlsx/xlsx_comments.xlsx", package = "SLutils")
XLSXsheets(xlsx)
}
