% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{XLSXwrite}
\alias{XLSXwrite}
\title{Write a XLSX file}
\usage{
XLSXwrite(dat, file, comments = NULL, commentsAuthor = NULL,
  base64 = FALSE, input = c("dataframe", "JSON"))
}
\arguments{
\item{dat}{dataframe or JSON string}

\item{file}{name of the output xlsx file}

\item{comments}{optional dataframe or JSON string}

\item{commentsAuthor}{option string}

\item{base64}{logical, whether to get the file as a base64 string}

\item{input}{\code{"dataframe"} if \code{dat} and \code{comments} are given as dataframes, \code{"JSON"} if they are given as JSON strings}
}
\value{
No value returned if \code{base64=FALSE}, otherwise a base64 string.
}
\description{
Write a XLSX file
}
\examples{
\dontrun{
dat <- data.frame(x=1:2, y=3:4)
comments <- data.frame(x=c(NA,NA), y=c(NA, "hello!"))
XLSXwrite(dat, "myfile.xlsx", comments)}
}
