% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx0.R
\name{XLSXwrite0}
\alias{XLSXwrite0}
\title{Write XLSX file}
\usage{
XLSXwrite0(..., file, sheetNames = NULL, author = "", title = "")
}
\arguments{
\item{...}{dataframes to be written (one dataframe per sheet)}

\item{file}{output xlsx file}

\item{sheetNames}{vector of sheet names; if \code{NULL}, default to \code{Sheet1}, \code{Sheet2}, ...}

\item{author}{author name}

\item{title}{title}
}
\description{
Write XLSX file
}
\note{
Current caveats:
\itemize{
     \item The number of digits is limited to 14.
}
}
\examples{
\dontrun{
XLSXwrite0(iris[1:20,], mtcars, file="test.xlsx",
          author="John Doe", title="MyTitle")
}
}
